Didn't really need kind=8 integers for this value; changed fortran version to kind=4. (=2 doesn't work). Reduces runtime to 0.594s

+++ 5/5.f90
@@ -1,7 +1,7 @@
 program problem5
 implicit none
 
-integer (kind=8) :: x
+integer (kind=4) :: x
 logical :: check
 x = 20*19
 do while (check(x).eqv..false.)
@@ -13,7 +13,7 @@ end program
 
 logical function check (n)
   implicit none
-  integer (kind=8) :: n,d
+  integer (kind=4) :: n,d
   do d=1,20,1
       if (mod(n,d).ne.0) then
           check = .false.
